create locale settings aware trait
