+++ config/vufind/Notices.yaml
@@ -22,12 +22,12 @@ styles:
 #             by adding <?=$this->notices()->renderNoticesForPosition('<position>')?> to your templates.
 notices:
 # Example 1
-  - style: warning
-    content: The library is closed today!
+#  - style: warning
+#    content: The library is closed today!
 # Example 2 (Notices will only be displayed for languages de, es and en. For es it will fall back to en)
-  - style: success
-    position: header
-    translations:
-      de: Willkommen in unserem Katalog!
-      es:
-      en: Welcome to our catalog!
+#  - style: success
+#    position: header
+#    translations:
+#      de: Willkommen in unserem Katalog!
+#      es:
+#      en: Welcome to our catalog!
