+++ module/VuFind/src/VuFind/I18n/Locale/LocaleSettingsAwareInterface.php
@@ -1,0 +1,54 @@
+<?php
+
+/**
+ * Locale settings aware marker interface.
+ *
+ * PHP version 8
+ *
+ * Copyright (C) Hebis Verbundzentrale 2026.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see
+ * <https://www.gnu.org/licenses/>.
+ *
+ * @category VuFind
+ * @package  I18n\Locale
+ * @author   Thomas Wagener <wagener@hebis.uni-frankfurt.de>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org Main Site
+ */
+
+namespace VuFind\I18n\Locale;
+
+use Laminas\Translator\TranslatorInterface;
+use VuFind\I18n\Translator\TranslatorAwareInterface;
+
+/**
+ * Locale settings aware marker interface.
+ *
+ * @category VuFind
+ * @package  I18n\Locale
+ * @author   Thomas Wagener <wagener@hebis.uni-frankfurt.de>
+ * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
+ * @link     https://vufind.org Main Site
+ */
+interface LocaleSettingsAwareInterface
+{
+    /**
+     * Set locale settings
+     *
+     * @param LocaleSettings $localeSettings Locale Settings
+     *
+     * @return LocaleSettingsAwareInterface
+     */
+    public function setLocaleSettings(LocaleSettings $localeSettings): LocaleSettingsAwareInterface;
+}
